% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchitectureModel.R
\name{annotatearchitecture}
\alias{annotatearchitecture}
\title{Annotate SNPs in a genetic architecture according to a model}
\usage{
annotatearchitecture(dat, s, thresholds = c(2.5, 3), sigma = NULL)
}
\arguments{
\item{dat}{data frame with the first two columns being "f" and "beta"}

\item{s}{The architecture model value for s}

\item{thresholds}{(Default: c(2.5,3)) The architecture thresholds to be used in s.d. units.}

\item{sigma}{(Default: NULL) The sigma to use for the data. Default: use that estimated from the data.}
}
\value{
The class of each datapoint
}
\description{
Compute the effect size prediction for a specific genomic architecture model, thereby annotating SNPs into classes according to how they fit with respect to the given thresholds.

Recall that we are not expecting individual SNPs to match their prediction, only the variance conditional on MAF to do so.
}
