% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchitecturePlot.R
\name{architectureplot}
\alias{architectureplot}
\title{Plot data annotated}
\usage{
architectureplot(
  dat,
  res = NULL,
  mycols = c("88888844", "#0000FF44", "#FF000044"),
  xlim = c(0.01, 0.5),
  ylim = NULL,
  xlab = "MAF",
  ylab = "Effect magnitude",
  log = "",
  tx = seq(0.005, 0.5, length.out = 100),
  main = "Genomic Architecture structure",
  bins.lty = 2,
  nbinom = 100,
  ...
)
}
\arguments{
\item{dat}{Data frame containing MAF and beta in first two columns.}

\item{res}{(Default: NULL)}

\item{mycols}{(Default: c(grey,blue,red)) three colours, one for each class of SNP}

\item{xlim}{(Default: c(0.01,0.5)) As provided to \code{plot}.}

\item{ylim}{(Default: NA, meaning use c(0,max(dat[,2])) As provided to \code{plot}.}

\item{xlab}{(Default: "MAF") As provided to \code{plot}.}

\item{ylab}{(Default: "Effect magnitude") As provided to \code{plot}.}

\item{log}{(Default: "") As provided to \code{plot}.}

\item{tx}{(Default: seq(0.005,0.5,length.out=100)) x values to evaluate the architecture curves at}

\item{main}{(Default: "Genomic Architecture structure") as provided to plot.}

\item{bins.lty}{(Default: 2) line type of the bin boundaries (if res provided)_}

\item{...}{Additional parameters to \code{plot}.}
}
\value{
A dataframe the same shape as provided with only the top k SNPs.
}
\description{
Filter data to keep only the top k by proportion of variance explained, which must be provided in a column called 'propexplained'.
}
