% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchitectureModel.R
\name{defaultbins}
\alias{defaultbins}
\title{Default bin locations for general purpose usage}
\usage{
defaultbins()
}
\value{
a vector of bin separators
}
\description{
Histogram bins that should be ok for most use; these are:

seq(0.05,0.5,by=0.025)
}
