% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchitectureModel.R
\name{defaultthresholds}
\alias{defaultthresholds}
\title{Default thresholds}
\usage{
defaultthresholds()
}
\value{
A pair of threshold values, the first for data to be considered, the second to be predicted.
}
\description{
Thresholds that should be accurate if they can be used. These are:

c(2.5,3)

in standard deviation units.

If you don't have data that has this large a variance, you may choose a smaller
threshold pair; c(1,2) is about as small as you should go.
}
