% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchitectureModel.R
\name{detail}
\alias{detail}
\title{Details of an architecture}
\usage{
detail(
  dat,
  s,
  thresholds = defaultthresholds(),
  bins = defaultbins(),
  range = c(-2, 2),
  range0 = c(-2, 2),
  sigma = NULL
)
}
\arguments{
\item{dat}{A data frame of (f,beta) pairs}

\item{s}{The architecture parameter to use}

\item{thresholds}{(Default :\code{\link{defaultthresholds}}) the tail areas to compare (in s.d units).}

\item{bins}{(Default: \code{\link{defaultbins}}) the histogram bins to use}

\item{range}{(Default: c(-2,2) The range of s that could be considered with finite values (if known).}

\item{range0}{(Default: c(-2,2) The range of s that was chosen by the user.}

\item{sigma}{(Default: NULL, meaning use the best value) Genomic architecture sigma. Don't change unless you know what you are doing...}
}
\description{
Details of an architecture, which may be inferred or defined manually.
}
\seealso{
\code{\link{architectureplot}} for visualisation.
}
