% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchitectureModel.R
\name{getbeta}
\alias{getbeta}
\title{Effect size of a SNP}
\usage{
getbeta(f, sigma, s)
}
\arguments{
\item{f}{The freqeuncy of the SNP}

\item{sigma}{The standard deviation}

\item{s}{The Selection parameter}
}
\value{
The variance explained
}
\description{
Get the mean effect of a genomic architecture model, i.e.
sigma (f (1-f))^(s/2)
}
