% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchitectureModel.R
\name{getsigma}
\alias{getsigma}
\title{Get the best estimate of sigma conditional on s}
\usage{
getsigma(ratio, thresholds)
}
\arguments{
\item{ratio}{The ratio of beta/betahat(f,s,sigma=1)}

\item{thresholds}{The architecture thresholds to be used in s.d. units.}
}
\value{
The estimate of sigma
}
\description{
Estimate sigma using the provided SNPs and a given s. This is done by ensuring that the whole distribution is scaled to obtain the theoretical proportion of SNPs above the second threshold, given the first, thereby ensuring that if s is correct then there should be no relationship between MAF and whether beta is over each threshold.

This uses a binary search of the possible thresholds that impact the choice, and is therefore logarithmic in the number of SNPs.
}
