% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchitectureModel.R
\name{lossfunction}
\alias{lossfunction}
\title{Compute the loss from summary statistics}
\usage{
lossfunction(ss)
}
\arguments{
\item{ss}{The ss as returned by \code{\link{lossfunctionss}}.}
}
\value{
The score of the summary statistics as a loss (higher is better)
}
\description{
Takes the loss function summary statistics table as returned from \code{\link{lossfunctionss}} and evaluates the loss, defined by the mean of the loss function for all SNPs across all MAF bins.
}
\seealso{
\code{\link{lossfunctionss}} to generate summary statistics, and \code{\link{fullsearch}} to search over s.
}
