% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchitectureModel.R
\name{lossfunctionss}
\alias{lossfunctionss}
\title{Calculation of histogram-based summary statistics of a distribution}
\usage{
lossfunctionss(
  dat,
  s,
  bins = defaultbins(),
  thresholds = defaultthresholds(),
  sigma = NULL
)
}
\arguments{
\item{dat}{A dataframe containing summary statistics from a GWAS/BayesS/LDAK analysis. In its first two columns we need: MAF (f) and beta.}

\item{s}{The selection coefficient assumed in the architecture model}

\item{bins}{(Default: bins=\code{\link{defaultbins}} The histogram bin boundaries assumed for creating summary statistcs, which need to be large enough to contain enough samples and small enough to retain resolution.}

\item{thresholds}{(Default: \code{\link{defaultthresholds}}) Standard deviation units for the summary statistic generation. Only SNPs above the first are used for model learning. The proportion above the second is predicted from the number above the first.}

\item{sigma}{(Default: NULL meaning estimate from data) The scale parameter of the Genomic Architecture. Do not provide this with any search procedure as the results can be poor.}
}
\value{
A data frame containing: bins, count0 (The number of SNPs in the bin), count (the number of SNPs above the first threshold), count2 (the number of SNPs above the second threshold)
}
\description{
Compute summary statistics from a dataset of MAF,beta pairs, by binning data into histograms and comparing whether the theoretical proportion of SNPs above a threshold matches the observed proportion.

Uses \code{\link{annotatearchitecture}} to annotate each SNP by with its prediction, and \code{\link{getsigma}} to estimate sigma given s and the dataset.
}
\seealso{
\code{\link{lossfunction}} to turn this summary into a loss, and \code{\link{fullsearch}} to search over s.
}
