% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchitectureSimulate.R
\name{make_test_data}
\alias{make_test_data}
\title{Simulate from a Genomic Architecture}
\usage{
make_test_data(N, sigma = 1, s = -1, frange = c(0.01, 0.5))
}
\arguments{
\item{N}{Number of SNPs to sample}

\item{sigma}{(default: 1) Baseline variance}

\item{s}{(default: -1) Genomic archictecture shape}

\item{frange}{(default: 0.01,0.5) range of frequencies to simulate for f}
}
\value{
A data frame containing N rows and four columns: f, beta, varexplained (variance explained) and propexplained (proportion of variance explained).
}
\description{
Simulate a genomic architecture from:

sigma^2 (f (1-f))^(s)

Where:
\code{f=unif(frange[1],frange[2])} and \code{beta=N(0,sigma*(f*(1-f))^(s/2))}.
}
