% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchitectureSimulate.R
\name{random_data}
\alias{random_data}
\title{Choose k random SNPs}
\usage{
random_data(testdata, k = 100)
}
\arguments{
\item{testdata}{The data frame containing frequency in the first column.}

\item{k}{The number of SNPs to retain}
}
\value{
A dataframe the same shape as provided with only min(k,dim(testdata)[1]) SNPs.
}
\description{
Keep at most k random SNPs from a dataframe.
}
