% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchitectureSimulate.R
\name{thresh_data}
\alias{thresh_data}
\title{Threshold a dataset based on frequency}
\usage{
thresh_data(testdata, t)
}
\arguments{
\item{testdata}{The data frame containing frequency in the first column.}

\item{t}{The desired threshold}
}
\value{
A dataframe the same shape as provided with only common SNPs retained.
}
\description{
Assuming the first column is frequency, return only rows where f>=t.
}
