% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchitectureSimulate.R
\name{top_data}
\alias{top_data}
\title{Retain only top SNPs by variance explained}
\usage{
top_data(testdata, k = 100)
}
\arguments{
\item{testdata}{The data frame containing frequency in the first column.}

\item{k}{The number of SNPs to retain}
}
\value{
A dataframe the same shape as provided with only the top k SNPs.
}
\description{
Filter data to keep only the top k by proportion of variance explained, which must be provided in a column called 'propexplained'.
}
