% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchitecturePlot.R
\name{architectureplot}
\alias{architectureplot}
\title{Plot data annotated}
\usage{
architectureplot(
  dat,
  res = NULL,
  mycols = c("#AAAAAA44", "#0000FF44", "#FF000044"),
  mar = c(6, 5, 3, 5),
  xlim = c(0.01, 0.5),
  ylim = NULL,
  xlab = "MAF",
  ylab = "Effect magnitude",
  log = "",
  tx = seq(0.005, 0.5, length.out = 100),
  main = "Genomic Architecture structure",
  show.data = TRUE,
  show.bins = TRUE,
  show.res = TRUE,
  data.pch = 19,
  data.cex = 0.5,
  bins.lty = 2,
  res.scale = 1,
  res.lwd = 1,
  res.col = "#008800FF",
  res.shade = "#88FF8844",
  res.lab = "Extreme class proportion",
  res.lab.line = 3,
  res.adj = 0.5,
  ...
)
}
\arguments{
\item{dat}{Data frame containing MAF and beta in first two columns.}

\item{res}{(Default: NULL)}

\item{mycols}{(Default: c(grey,blue,red)) three colours, one for each class of SNP}

\item{mar}{(Default: c(6,5,3,5)) Margins, passed to par(mar=mar) if not NA.}

\item{xlim}{(Default: c(0.01,0.5)) As provided to \code{plot}.}

\item{ylim}{(Default: NA, meaning use c(0,max(dat[,2])) As provided to \code{plot}.}

\item{xlab}{(Default: "MAF") As provided to \code{plot}.}

\item{ylab}{(Default: "Effect magnitude") As provided to \code{plot}.}

\item{log}{(Default: "") As provided to \code{plot}.}

\item{tx}{(Default: seq(0.005,0.5,length.out=100)) x values to evaluate the architecture curves at}

\item{main}{(Default: "Genomic Architecture structure") as provided to plot.}

\item{show.data}{(Default: TRUE) whether to plot the data.}

\item{show.bins}{(Default: TRUE) whether to plot the bins, if res provided.}

\item{show.res}{(Default: TRUE) whether to show the result proportions, if res provided.}

\item{data.pch}{(Default: 19) pch for data.}

\item{data.cex}{(Default: 0.5) cex for data.}

\item{bins.lty}{(Default: 2) line type of the bin boundaries (if res provided).}

\item{res.scale}{(Default: 1) vertical scaling of the results. Set to values < 1 to restrict the class proportion data into the bottom segment, which is often excluded from architecture plots.}

\item{res.lwd}{(Default: 1) Line width of the class proportion results.}

\item{res.col}{(Default: "#008800FF" which is dark green) colour for all results to be displayed.}

\item{res.shade}{(Default: "#88FF8844" which is pale green) colour for the theoretically expected variation in proportions per frequency bin.}

\item{res.lab}{(Default: "Extreme class proportion") axis label for the results. Written in res.col.}

\item{res.lab.line}{(Default: 3) line to write res.lab with mtext.}

\item{res.adj}{(Default:0.5) vertical location to write res.lab. Try setting to 0 if you have used res.scale<1.}

\item{...}{Additional parameters to \code{plot}.}
}
\value{
A dataframe the same shape as provided with only the top k SNPs.
}
\description{
Filter data to keep only the top k by proportion of variance explained, which must be provided in a column called 'propexplained'.
}
