% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchitectureModel.R
\name{bootstrap}
\alias{bootstrap}
\title{Obtain a bootstrap sample of Genomic Architecture s}
\usage{
bootstrap(
  nbs,
  dat,
  bins = seq(0.05, 0.5, by = 0.025),
  thresholds = c(2.5, 3),
  sigma = NULL,
  verbose = TRUE,
  s = 0,
  ...
)
}
\arguments{
\item{nbs}{The number of bootstrap samples to obtain. We recommend 20 for exploration, 100 for estimating standard deviations and more for estimating the shape of the distribution.}

\item{dat}{A data frame of (f,beta) pairs}

\item{bins}{(default: seq(0.05,0.5,by=0.025)) the histogram bins to use}

\item{thresholds}{(Default: c(2.5,3)) the tail areas to compare (in s.d units).}

\item{sigma}{(Default: NULL, meaning use the best value) Genomic architecture sigma}

\item{verbose}{(Default: TRUE) Whether to print iteration progress.}

\item{s}{(Default: 0) Initial value of the search}

\item{...}{Additional parameters to \code{\link[stats]{optim}}.}
}
\value{
A vector of length \code{nbs} of estimates of \code{s}.
}
\description{
Resample the data \code{dat} with replacement a given number \code{nbs} times, for each running \code{\link{fullsearch}} and reporting the inferred parameter.
}
\seealso{
\code{\link{linesearch}} to explore a range of s, or \code{\link{\code{\link{fullsearch}} }} to search a single dataset.
}
