% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchitectureModel.R
\name{fullsearch}
\alias{fullsearch}
\title{Search s for the best parameter}
\usage{
fullsearch(
  dat,
  bins = seq(0.05, 0.5, by = 0.025),
  thresholds = c(2.5, 3),
  s = 0,
  sigma = NULL,
  range = c(-2, 2),
  ...
)
}
\arguments{
\item{dat}{A data frame of (f,beta) pairs}

\item{bins}{(Default: seq(0.05,0.5,by=0.025)) the histogram bins to use}

\item{thresholds}{(Default: c(2.5,3)) the tail areas to compare (in s.d units).}

\item{s}{(Default: 0) Initial value of the search}

\item{sigma}{(Default: NULL, meaning use the best value) Genomic architecture sigma}

\item{range}{(Default: c(-2,2)) The lower and upper bounds provided to}

\item{...}{Additional parameters to \code{\link[stats]{optim}}. Note: You cannot provide  \code{control} as this is already provided.}
}
\value{
A list as returned by \code{\link[stats]{optim}}, with additionally \code{dat}, \code{thresholds}, \code{bins}, \code{range} and \code{sigma} as provided, and \code{s} as inferred.
}
\description{
Optimize \code{\link{lossfunctionp}} using  \code{\link[stats]{optim}} with \code{method="Brent"}.
}
\seealso{
\code{\link{linesearch}} to explore a range of s, or \code{\link{bootstrap}} to generate a distribution of estimates of s.
}
